ajuste na propriedade 'Saldo', alterada para 'decimal'

+++ SysContabil/src/Dominio/Dominio/Entidades/Balancete.cs
@@ -6,7 +6,7 @@ namespace Dominio.Entidades
 {
     public class Balancete
     {
-        public Balancete(string nomeDaConta, string numeroDaConta, double saldo)
+        public Balancete(string nomeDaConta, string numeroDaConta, decimal saldo)
         {
             NomeDaConta = nomeDaConta;
             NumeroDaConta = numeroDaConta;
@@ -16,7 +16,7 @@ namespace Dominio.Entidades
         public int Id { get; private set; }
         public string NomeDaConta { get; set; }
         public string NumeroDaConta { get; set; }
-        public double Saldo { get; set; }
+        public decimal Saldo { get; set; }
 
         public IEnumerable<PlanoDeConta> PlanoDeContas { get; private set; }
